requirements.txt
bcrypt==3.2.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
cssutils==2.5.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Login==0.6.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
premailer==3.10.0
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.11
uuid7==0.1.0
Werkzeug==2.2.2
WTForms==3.0.1
yagmail==0.15.283
zipp==3.8.1
